refactor: update component imports and add new icons
