+++ client/src/components/PostFooter/PostFooter.jsx
@@ -1,116 +1,16 @@
 import Button from "../Button.jsx";
 import style from "./PostFooter.module.css";
 import PropTypes from "prop-types";
-
-const Like = ({ fill = false }) => {
-  return (
-    <svg
-      className={style.icon}
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g id="Interface / Heart_01">
-        <path
-          id="Vector"
-          d="M12 7.69431C10 2.99988 3 3.49988 3 9.49991C3 15.4999 12 20.5001 12 20.5001C12 20.5001 21 15.4999 21 9.49991C21 3.49988 14 2.99988 12 7.69431Z"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          fill={fill ? "currentColor" : "none"}
-        />
-      </g>
-    </svg>
-  );
-};
-const Comment = () => {
-  return (
-    <svg
-      className={style.icon}
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g id="Communication / Chat">
-        <path
-          id="Vector"
-          d="M5.59961 19.9203L7.12357 18.7012L7.13478 18.6926C7.45249 18.4384 7.61281 18.3101 7.79168 18.2188C7.95216 18.1368 8.12328 18.0771 8.2998 18.0408C8.49877 18 8.70603 18 9.12207 18H17.8031C18.921 18 19.4806 18 19.908 17.7822C20.2843 17.5905 20.5905 17.2842 20.7822 16.9079C21 16.4805 21 15.9215 21 14.8036V7.19691C21 6.07899 21 5.5192 20.7822 5.0918C20.5905 4.71547 20.2837 4.40973 19.9074 4.21799C19.4796 4 18.9203 4 17.8002 4H6.2002C5.08009 4 4.51962 4 4.0918 4.21799C3.71547 4.40973 3.40973 4.71547 3.21799 5.0918C3 5.51962 3 6.08009 3 7.2002V18.6712C3 19.7369 3 20.2696 3.21846 20.5433C3.40845 20.7813 3.69644 20.9198 4.00098 20.9195C4.35115 20.9191 4.76744 20.5861 5.59961 19.9203Z"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </g>
-    </svg>
-  );
-};
-const Share = () => {
-  return (
-    <svg
-      className={style.icon}
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g id="Communication / Share_Android">
-        <path
-          id="Vector"
-          d="M9 13.5L15 16.5M15 7.5L9 10.5M18 21C16.3431 21 15 19.6569 15 18C15 16.3431 16.3431 15 18 15C19.6569 15 21 16.3431 21 18C21 19.6569 19.6569 21 18 21ZM6 15C4.34315 15 3 13.6569 3 12C3 10.3431 4.34315 9 6 9C7.65685 9 9 10.3431 9 12C9 13.6569 7.65685 15 6 15ZM18 9C16.3431 9 15 7.65685 15 6C15 4.34315 16.3431 3 18 3C19.6569 3 21 4.34315 21 6C21 7.65685 19.6569 9 18 9Z"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </g>
-    </svg>
-  );
-};
-const More = () => {
-  return (
-    <svg
-      className={style.icon}
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <g id="Menu / More_Horizontal">
-        <g id="Vector">
-          <path
-            d="M17 12C17 12.5523 17.4477 13 18 13C18.5523 13 19 12.5523 19 12C19 11.4477 18.5523 11 18 11C17.4477 11 17 11.4477 17 12Z"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M11 12C11 12.5523 11.4477 13 12 13C12.5523 13 13 12.5523 13 12C13 11.4477 12.5523 11 12 11C11.4477 11 11 11.4477 11 12Z"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M5 12C5 12.5523 5.44772 13 6 13C6.55228 13 7 12.5523 7 12C7 11.4477 6.55228 11 6 11C5.44772 11 5 11.4477 5 12Z"
-            stroke="currentColor"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </g>
-      </g>
-    </svg>
-  );
-};
+import { LikeIcon, CommentIcon, ShareIcon, MoreIcon } from "../icons/index.js";
 
 function PostFooter({ tags }) {
   return (
     <footer className={style.footer}>
       <section className={style.wrapper}>
         <section className={style.actionsContainer}>
-          <Button icon={<Like fill={false} />} />
-          <Button icon={<Comment />} />
-          <Button icon={<Share />} />
+          <Button icon={<LikeIcon style={style.icon} fill={false} />} />
+          <Button icon={<CommentIcon style={style.icon} />} />
+          <Button icon={<ShareIcon style={style.icon} />} />
         </section>
         <section className={style.tagsContainer}>
           <ul>
@@ -125,17 +25,13 @@ function PostFooter({ tags }) {
           </ul>
         </section>
       </section>
-      <Button icon={<More />} />
+      <Button icon={<MoreIcon style={style.icon} />} />
     </footer>
   );
 }
 
 PostFooter.propTypes = {
   tags: PropTypes.arrayOf(PropTypes.string).isRequired,
-};
-
-Like.propTypes = {
-  fill: PropTypes.bool,
 };
 
 export default PostFooter;
