+++ client/src/components/icons/index.js
@@ -1,3 +1,8 @@
 export { default as HomeIcon } from "./HomeIcon.jsx";
 export { default as SearchIcon } from "./SearchIcon.jsx";
 export { default as ProfileIcon } from "./ProfileIcon.jsx";
+export { default as GoogleIcon } from "./GoogleIcon.jsx";
+export { default as LikeIcon } from "./LikeIcon.jsx";
+export { default as CommentIcon } from "./CommentIcon.jsx";
+export { default as ShareIcon } from "./ShareIcon.jsx";
+export { default as MoreIcon } from "./MoreIcon.jsx";
