+++ client/src/index.css
@@ -72,7 +72,7 @@ button:disabled {
 body {
   font-family: var(--font-sans), sans-serif;
   font-weight: 500;
-  padding: 10rem 1rem 0 1rem;
+  padding: 0 1rem;
 }
 
 @media (max-width: 768px) {
