--- client/src/pages/Home.jsx
@@ -1,5 +1,0 @@
-function Home() {
-  return <>Home Page</>;
-}
-
-export default Home;
