+++ client/src/pages/Home/Home.jsx
@@ -1,0 +1,5 @@
+function Home() {
+  return <>Home Page</>;
+}
+
+export default Home;
