+++ client/src/pages/NewPost/NewPost.jsx
@@ -1,4 +1,4 @@
-import CreatePostForm from "../components/CreatePostForm/CreatePostForm.jsx";
+import CreatePostForm from "../../components/CreatePostForm/CreatePostForm.jsx";
 
 export default function NewPostPage() {
   return (
